+++ Core/src/ecs/EntityID.h
@@ -6,6 +6,6 @@ namespace tryn::ecs
 	struct EntityID
 	{
 		uint32_t ID = 0;
-		uint16_t archetype = 0;
+		uint32_t archetype = 0;
 	};
 }
